Move MSSQL custom types to a `mssql_dialect` file
We also create our own dialect subclass where we can apply
cusomtisations.
